feat: add new API routes for managing output cleanup items and library links
- Implemented `latest-json.js` to fetch the latest updater manifest from GitHub.
- Created `library-link.js` for submitting library link requests to GitHub issues.
- Added `library-links-data.js` to fetch library links data from the repository.
- Developed `livery-patterns-data.js` to retrieve livery patterns data from GitHub.
- Introduced `output-cleanup-item-submission.js` for submitting output cleanup items.
- Created `output-cleanup-items-data.js` to fetch output cleanup items data.
- Added `release-redirect.js` for redirecting to GitHub release pages.
- Implemented `update-release.js` to fetch and normalize release data.
- Added `OutputCleanup.vue` component for managing output cleanup items in the UI.
